Move Store to a dedicated package for better isolation
