+++ protoarray/src/test/java/tech/pegasys/teku/protoarray/NoVotesTest.java
@@ -29,7 +29,7 @@ import java.util.List;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.storage.api.StubStorageUpdateChannel;
-import tech.pegasys.teku.storage.client.Store;
+import tech.pegasys.teku.storage.store.Store;
 
 public class NoVotesTest {
