+++ protoarray/src/test/java/tech/pegasys/teku/protoarray/ProtoArrayForkChoiceStrategyTest.java
@@ -32,7 +32,7 @@ import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.datastructures.forkchoice.MutableStore;
 import tech.pegasys.teku.datastructures.forkchoice.VoteTracker;
 import tech.pegasys.teku.storage.api.StubStorageUpdateChannel;
-import tech.pegasys.teku.storage.client.Store;
+import tech.pegasys.teku.storage.store.Store;
 
 public class ProtoArrayForkChoiceStrategyTest {
