+++ storage/src/main/java/tech/pegasys/teku/storage/api/StorageUpdateChannel.java
@@ -14,8 +14,8 @@
 package tech.pegasys.teku.storage.api;
 
 import java.util.Optional;
-import tech.pegasys.teku.storage.client.Store;
 import tech.pegasys.teku.storage.events.StorageUpdate;
+import tech.pegasys.teku.storage.store.Store;
 import tech.pegasys.teku.util.async.SafeFuture;
 
 public interface StorageUpdateChannel {
