+++ storage/src/test/java/tech/pegasys/teku/storage/server/AbstractStorageBackedDatabaseTest.java
@@ -31,8 +31,8 @@ import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.Checkpoint;
 import tech.pegasys.teku.storage.api.DatabaseBackedStorageUpdateChannel;
-import tech.pegasys.teku.storage.client.Store;
-import tech.pegasys.teku.storage.client.Store.Transaction;
+import tech.pegasys.teku.storage.store.Store;
+import tech.pegasys.teku.storage.store.Store.Transaction;
 import tech.pegasys.teku.util.config.StateStorageMode;
 import tech.pegasys.teku.util.file.FileUtil;
