+++ storage/src/testFixtures/java/tech/pegasys/teku/storage/api/DatabaseBackedStorageUpdateChannel.java
@@ -14,9 +14,9 @@
 package tech.pegasys.teku.storage.api;
 
 import java.util.Optional;
-import tech.pegasys.teku.storage.client.Store;
 import tech.pegasys.teku.storage.events.StorageUpdate;
 import tech.pegasys.teku.storage.server.Database;
+import tech.pegasys.teku.storage.store.Store;
 import tech.pegasys.teku.util.async.SafeFuture;
 
 public class DatabaseBackedStorageUpdateChannel implements StorageUpdateChannel {
